multi_agent_chatbot/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0

# LlamaIndex dependencies
llama-index>=0.9.0
llama-index-llms-azure-openai>=0.1.0
llama-index-embeddings-azure-openai>=0.1.0

# Additional dependencies
tiktoken>=0.5.0
